finance-agent/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pypdf
faiss-cpu
numpy
python-dotenv
google-genai
python-docx
streamlit